다이어리 삭제, 삭제 팝업 구현

+++ lib/feature/diary/presentation/viewmodels/new_diary_state.dart
@@ -67,7 +67,7 @@ abstract class NewDiaryState with _$NewDiaryState {
       type: type,
       title: title.isNotEmpty ? title : null,
       content: content.isNotEmpty ? content : null,
-      rating: type == 'REVIEW' ? rating.toInt() : null,
+      rating: type == 'REVIEW' ? rating.toDouble() : null,
       img: type == 'PHOTO' ? imgUrl : null,
       cost: type == 'MONEY' ? cost : null,
       isPublic: isPublic,
